external.py: rename existing external.py to s3.py
